Refactor DashboardPage to Server Component with DashboardClient
